Add detailed list of inappropriate behaviors from Mozilla CPG.
On 2019/Jun/12, the GNOME online Code of Conduct committee requested
that the Code of Conduct include the more detailed list of inappropriate
behaviors. They were notified that would make the Code of Conduct need
to be licensed under a CC-BY-SA license, and they said it was fine.

Add additional attribution for the Mozilla Community Participation
Guidelines, and one of the sources it derives from. Update license to
Creative Commons Attribution Share-Alike 3.0.