Change data retention wording to focus on inappropriate behavior, not labels.
The term "repeat offenders" is an unclear label.

Clarify that what the committee is recording is a person's inappropriate
behavior, or their borderline inappropriate behavior. This is then used
to identify when the behavior (or a related border-line inappropriate
behavior) has been repeated.

The phrase "Create a dataset to base future enforcement actions on" may
trigger a lot of concern in people, especially in our age of
surveillance culture.

Clarify that past records may be used to determine whether a stricter
consequence is needed in response to the most recent report.