Add exception to record removal in the case of community safety issues.
There may be cases where a person has not been sanctioned, but it's
important to retain a record of an incident where a community member's
safety was at risk.

For example, some reporters may fear retribution from the reported
person. This is more likely to happen in cases of sexual harassment,
which represents a risk to the safety of GNOME community members.

The reporter may not want the committee to take action unless another
report comes in about the same reported person. In this case, the
committee may not be able to act on the report immediately.

However, the report is useful in establishing a pattern of inappropriate
behavior. It may be used to determine whether a more strict consequence
is needed if an incident report involving the same reported person is
filed.