Add guide for GNOME moderator teams.
GNOME has moderator teams for individual communities. They may need to
be involved in communicating with the GNOME online Code of Conduct
committee. Sometimes they will need to take direct action in safe and
simple incidents. Other times they'll need to defer to the committee.
Make it clear which types of incidents fall into which categories.