Integrate details about public reports and employer reports.
Otter Tech is not a legal advisor and does not provide legal advice.
Otter Tech is not a public relations advisor and does not provide public
relations advice.

The following text was integrated from the committee private procedures.
It was modified so examples cover both events and online incidents.
Language was changed to use reporter, reported person, witness, and
impacted people.