Add respecting technology choices to community guidelines.
The GNOME board was concerned that being harassed for using a specific
tool (e.g. a proprietary tool) was not covered by the CoC. 'Choice of
technology' should never be added to a list of protected classes, but we
can emphasize that it should be respected in the community guidelines.