Cover stalking through private messaging.
The GNOME board was concerned that the scope of the document doesn't
cover private communication between GNOME participants. For example, a
GNOME participant could receive harassing or stalking messages via SMS,
which is not covered under the current scope.

Restrict private communication to be in scope if a participant is being
directly stalked or harassed. Otherwise private messages should remain
out of scope.